+++ kmath-core/src/commonMain/kotlin/kscience/kmath/expressions/Expression.kt
@@ -56,21 +56,6 @@ public operator fun <T> Expression<T>.invoke(vararg pairs: Pair<Symbol, T>): T =
 public operator fun <T> Expression<T>.invoke(vararg pairs: Pair<String, T>): T =
     invoke(mapOf(*pairs).mapKeys { StringSymbol(it.key) })
 
-/**
- * And object that could be differentiated
- */
-public interface Differentiable<T> {
-    public fun derivative(orders: Map<Symbol, Int>): T
-}
-
-public interface DifferentiableExpression<T> : Differentiable<Expression<T>>, Expression<T>
-
-public fun <T> DifferentiableExpression<T>.derivative(vararg orders: Pair<Symbol, Int>): Expression<T> =
-    derivative(mapOf(*orders))
-
-public fun <T> DifferentiableExpression<T>.derivative(symbol: Symbol): Expression<T> = derivative(symbol to 1)
-
-public fun <T> DifferentiableExpression<T>.derivative(name: String): Expression<T> = derivative(StringSymbol(name) to 1)
 
 /**
  * A context for expression construction
@@ -95,6 +80,10 @@ public interface ExpressionAlgebra<in T, E> : Algebra<E> {
      */
     public fun const(value: T): E
 }
+
+//public interface ExpressionBuilder<T, E, A : ExpressionAlgebra<T, E>> {
+//    public fun expression(block: A.() -> E): Expression<T>
+//}
 
 /**
  * Bind a given [Symbol] to this context variable and produce context-specific object.
